requirements.txt
Flask
schedule
SQLAlchemy
rich
python-dotenv
pyjwt